Run dialog parser in runtime
