+++ docs/OverlaySpec.md
@@ -1,0 +1,23 @@
+# Overlay Spec
+
+**What are the things that can appear on top of the normal content?**
+
+- Toast
+- Modal
+- Dialog
+  - Choices
+- Backstage
+  - Preferences
+  - Optional: Dialog??
+
+**Which one appears on top?**
+
+- Toast
+- Modal
+  - blocks access on anything below
+- Backstage
+  - blocks access on anything below
+- Dialog: choices
+  - blocks access on anything below
+- Dialog
+- Everything else
