+++ src/routes/(pineapple)/pineapple/+page.svelte
@@ -2,6 +2,9 @@
 	import { PUBLIC_CRINGE_USERNAME } from "$env/static/public";
 	import { showComponentInToast, showTextInToast } from "$pkg/components/pineapple/toast/Toast";
 	import TestCard from "$pkg/components/pineapple/toast/custom-toast/TestCustomToast.svelte";
+	import TestDialogYarn from "./TestDialog.yarn?raw";
+	import { parseYarn } from "$pkg/scripts/pineapple_fiber/PineappleFiberParser";
+	import { dialogManager } from "$pkg";
 
 	let testingQueueNumber = 1;
 	const testingRandomPhrases = [
@@ -9,6 +12,11 @@
 		"Niko the Baikal seal",
 		"Niko the Baikal seal\nfrom Toba Aquarium"
 	];
+
+	const dialogTree = parseYarn(TestDialogYarn);
+	// todo: parse the result
+	// todo: run the interpreter when needed
+	// todo: do parsing in dialogManager
 </script>
 
 <svelte:head>
@@ -21,7 +29,6 @@
 </svelte:head>
 
 <div class="card default-card">
-
 	<button
 		class="btn variant-filled-secondary"
 		on:click={() => {
@@ -33,6 +40,12 @@
 			showTextInToast(`${testingQueueNumber} ${testingRandomPhrases[testingQueueNumber]}`);
 			testingQueueNumber = (testingQueueNumber + 1) % testingRandomPhrases.length;
 		}}><h3>Handy toast</h3></button>
+	<button
+		class="btn variant-filled-secondary"
+		on:click={() => {
+				dialogManager.setDialogTree(dialogTree);
+				dialogManager.toggleDialogOverlay();
+		}}><h3>Test dialog</h3></button>
 </div>
 
 <style lang="postcss">
