+++ src/routes/(pineapple)/pineapple/TestDialog.yarn
@@ -1,0 +1,7 @@
+title: ErrorStart
+position: 0,0
+---
+Portrait: AresDisappointed
+<p>Huh... We didn't find that page. Sorry about that. You can go back to our <a href="/">homepage.</a><br>
+Gomen oomf chan o(TヘTo)</p>
+===
