+++ vite.config.ts
@@ -2,5 +2,6 @@ import { sveltekit } from "@sveltejs/kit/vite";
 import { defineConfig } from "vite";
 
 export default defineConfig({
-	plugins: [sveltekit()]
+	plugins: [sveltekit()],
+	assetsInclude: ['**/*.yarn']
 });
